+++ src/Core/global_functions.php
@@ -20,58 +20,6 @@
  * @license     http://www.gnu.org/licenses/agpl.txt AGPL License
  * @link        http://git.novatlantis.it Nova Atlantis Ltd
  */
-
-if (!function_exists('compressionFromFile')) {
-    /**
-     * Returns the compression from a file
-     * @param string $filename Filename
-     * @return string|false
-     */
-    function compressionFromFile($filename)
-    {
-        //Gets extension
-        $extension = extensionFromFile($filename);
-
-        $compressions = ['sql.bz2' => 'bzip2', 'sql.gz' => 'gzip', 'sql' => false];
-
-        if (!array_key_exists($extension, $compressions)) {
-            return null;
-        }
-
-        return $compressions[$extension];
-    }
-}
-
-if (!function_exists('extensionFromCompression')) {
-    /**
-     * Returns the extension from a compression
-     * @param string $compression Compression
-     * @return string|bool
-     */
-    function extensionFromCompression($compression)
-    {
-        $compressions = ['sql.bz2' => 'bzip2', 'sql.gz' => 'gzip', 'sql' => false];
-
-        return array_search($compression, $compressions, true);
-    }
-}
-
-if (!function_exists('extensionFromFile')) {
-    /**
-     * Returns the extension from a file
-     * @param string $filename Filename
-     * @return string|null
-     */
-    function extensionFromFile($filename)
-    {
-        if (!preg_match('/\.(sql(\.(gz|bz2))?)$/', $filename, $matches) || empty($matches[1])) {
-            return null;
-        }
-
-        return $matches[1];
-    }
-}
-
 if (!function_exists('isPositive')) {
     /**
      * Checks if a string is a positive number