+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 * added `BackupExport::send()` to set up the email recipient, so by calling the
     `export()` method the backup file will be sent via email;
 * added `BackupTrait`,  that provides some methods used by all other classes;
+* global functions `compressionFromFile()`, `extensionFromCompression()` and
+    `extensionFromFile()` have been replaced with the `getCompression()` and
+    `getExtension()` methods provided by the `BackupTrait`;
 * all methods belonging to the `BackupManager` class are no longer static.
 
 ## 1.0 branch